refactor: ajustes nas migrations e no banco de dados
