+++ Unihub.Infraestrutura/ConfiguracaoEntidades/AulasNaSemanaConfiguracao.cs
@@ -1,5 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Unihub.Dominio.Entidades;
 
 namespace Unihub.Infraestrutura.ConfiguracaoEntidades
