+++ Unihub.Infraestrutura/Unihub.Infraestrutura.csproj
@@ -25,7 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Migracoes\" />
+    <Folder Include="Migrations\" />
   </ItemGroup>
 
 </Project>
